test: exec
